FAKEAlert/requirements.txt
flask
flask-sqlalchemy
flask-login
requests
google-generativeai